Alright doing what needed to be done creating a Gravity Spy table class all on its own. This is now completed (in draft form) and the draft works
