+++ gwpy/table/__init__.py
@@ -33,6 +33,7 @@ tables of time-stamped GW (or other) events.
 # load tables
 from astropy.table import (Column, Table)
 from .table import (EventColumn, EventTable)
+from .gravityspy import GravitySpyTable
 
 # attach unified I/O
 from . import io
